Split search queries into different files
